+++ projects/ontimize-web-ngx/src/lib/components/o-service-base-component.class.ts
@@ -337,6 +337,11 @@ export abstract class AbstractOServiceBaseComponent<T extends AbstractComponentS
   }
 
 
+  public configureAdapter() {
+    this.queryArgumentAdapter = this.injector.get(OntimizeQueryArgumentsAdapter);
+  }
+
+
   configureService() {
     let configureServiceArgs: OConfigureServiceArgs = { injector: this.injector, baseService: OntimizeService, entity: this.entity, service: this.service, serviceType: this.serviceType }
     if (Util.isDefined(this.configureServiceArgs)) {